backend/apps/image-storage/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic == 2.*
cloudinary==1.36.0
python-multipart